+++ External/Include/Script/CInfectedStrong_A.h
@@ -31,7 +31,7 @@ public:
     void JpsAlgorithm(Int32 x, Int32 z);
     void SetDestPos(Vec3 _vPos);
    float GetHp() { return m_fHP; }
-    void SetHp(UINT _iHp) { m_fHP = _iHp; }
+   void SetHp(float _iHp) { m_fHP = _iHp; }
 private:
     bool Func(CGameObject* _lhs, CGameObject* _rhs);
     void ProcessEnemy();
